removed bootstrap and implemented tailwind + daisyui instead

+++ src/main/java/watchtogether/data/PlaylistRepository.java
@@ -3,5 +3,7 @@ package watchtogether.data;
 import org.springframework.data.jpa.repository.JpaRepository;
 import watchtogether.models.Playlist;
 
-public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {
+
+
+public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
 }
